+++ src/Cwaff-Passives/VolcanicAmmolet.cs
@@ -17,7 +17,7 @@ public class VolcanicAmmolet : CwaffBlankModificationItem, ICustomBlankDoer
     {
         PassiveItem item  = Lazy.SetupPassive<VolcanicAmmolet>(ItemName, ShortDescription, LongDescription, Lore);
         item.quality      = ItemQuality.B;
-        ItemBuilder.AddPassiveStatModifier(item, StatType.AdditionalBlanksPerFloor, 1f, StatModifier.ModifyMethod.ADDITIVE);
+        item.AddPassiveStatModifier(StatType.AdditionalBlanksPerFloor, 1f, StatModifier.ModifyMethod.ADDITIVE);
         item.AddToSubShop(ItemBuilder.ShopType.OldRed);
 
         _LargeVolcanicExplosion = Explosions.DefaultLarge.Clone();